requirements.txt
transformers 
sentence_transformers 
gradio
google-generativeai==0.3.1
langchain-google-genai
langchain